requirements.txt
pandas
scikit-learn
joblib
newspaper3k